api/requirements.txt
awslambdaric
boto3
openai
pinecone-client
protobuf
sentence-transformers
slack-sdk
torch